requirements.txt
antlr4-python3-runtime==4.9.3
bitarray==2.9.2
bitstring==4.2.3
contourpy==1.2.1
coverage==7.5.3
cycler==0.12.1
dill==0.3.8
fonttools==4.53.0
gdown==5.2.0
hydra-core==1.3.2
juliacall==0.9.20
juliapkg==0.1.13
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==1.26.4
omegaconf==2.3.0
optimparallel==0.1.3
packaging==24.1
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
scipy==1.14.0
semver==3.0.2
six==1.16.0
z3-solver==4.13.0.0
torch
yapf==0.32.0
docformatter==1.4
isort==5.10.1
colorlog==6.9.0